+++ src/components/Futa/ResizeableTable/ResizeableTableHeader/ResizeableTableHeader.module.css
@@ -1,0 +1,180 @@
+/* Container for the search and sort section */
+.header {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    align-items: center;
+    padding: 8px;
+    gap: 8px;
+    width: 100%;
+    height: auto;
+}
+
+/* Text search input and sort toggle */
+.search_and_sort {
+    flex: 1;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    gap: 4px;
+    height: 100%;
+}
+
+/* Wrapper for search input icon, text field, and clear button */
+.text_search_box {
+    flex: 1;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    width: 100%;
+    padding-left: 8px;
+    gap: 4px;
+    background-color: var(--dark2);
+    height: 100%;
+}
+
+.text_search_box > input {
+    width: 100%;
+    padding: 0 8px;
+    background: transparent;
+    color: var(--text1);
+    outline: none;
+    border: none;
+    opacity: 1;
+    transition: all var(--animation-speed) ease-in;
+    cursor: normal;
+    font-size: 12px;
+}
+
+/* Wrapper for toggles */
+.filters {
+    display: flex;
+    flex-direction: row;
+    gap: 8px;
+    height: 100%;
+}
+
+.filters > button {
+    min-width: 110px;
+    height: 25px;
+    border: 1px solid var(--dark3);
+    color: var(--text2);
+    background-color: var(--dark2);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    gap: 10px;
+    font-size: 12px;
+    cursor: pointer;
+    outline: none;
+}
+
+.filters > button:hover {
+    transition: all var(--animation-speed) ease-in-out;
+    color: var(--accent1);
+    border: 1px solid var(--accent1);
+}
+
+.button_on {
+    border: 1px solid var(--accent1) !important;
+}
+
+.button_on * {
+    color: var(--accent1);
+}
+
+.button_off {
+    border: 1px solid var(--dark3) !important;
+    color: var(--text2) !important;
+}
+
+.button_off * {
+    color: var(--text2) !important;
+}
+
+.sort_clickable {
+    position: relative;
+}
+
+.sort_selection {
+    display: flex;
+    gap: 4px;
+}
+
+.open_dropdown_clickable {
+    border-left: 4px solid var(--accent1, #62ebf1);
+    background-color: var(--dark2);
+    display: flex;
+    padding: 4px 4px 4px 8px;
+    justify-content: space-between;
+    align-items: center;
+    cursor: pointer;
+    width: 160px;
+    text-transform: uppercase;
+}
+
+.open_dropdown_clickable p {
+    font-size: 14px;
+    color: var(--accent1);
+}
+
+.sort_direction {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    cursor: pointer;
+}
+
+.dropdown {
+    height: auto;
+    width: 100%;
+    background-color: var(--dark2);
+    position: absolute;
+    top: 30px;
+    z-index: 2;
+    display: flex;
+    flex-direction: column;
+    gap: 2px;
+}
+
+.dropdown > p {
+    font-size: 12px;
+    width: 100%;
+    transition: all var(--animation-speed) ease-in-out;
+    cursor: pointer;
+    padding: 0 4px;
+    text-transform: uppercase;
+}
+
+.active_sort {
+    color: var(--dark2) !important;
+    background-color: var(--accent1);
+}
+
+.inactive_sort {
+    color: var(--accent1);
+}
+
+.inactive_sort:hover {
+    background: var(--text3);
+}
+
+/* Responsive styles */
+@media (max-width: 1024px) {
+    .header {
+        flex-direction: column;
+        padding-left: 0px;
+    }
+
+    .header > * {
+        display: flex;
+        width: 100%;
+        height: 29px !important;
+    }
+
+    .filters > * {
+        flex-grow: 1;
+    }
+}
