+++ src/components/Futa/SearchableTicker/SearchableTicker.tsx
@@ -1,10 +1,5 @@
-import { NumberSize, Resizable } from 're-resizable';
-import { Direction } from 're-resizable/lib/resizer';
 import { useContext, useEffect, useMemo, useRef, useState } from 'react';
-import { BiSearch } from 'react-icons/bi';
-import { FaEye } from 'react-icons/fa';
-import { IoIosArrowDown, IoIosArrowUp } from 'react-icons/io';
-import { MdClose } from 'react-icons/md';
+
 import { AuctionDataIF, diffHashSig } from '../../../ambient-utils/dataLayer';
 import { AuctionsContext } from '../../../contexts/AuctionsContext';
 import { FutaSearchableTickerContext } from '../../../contexts/Futa/FutaSearchableTickerContext';
@@ -14,16 +9,14 @@ import {
 } from '../../../pages/platformFuta/Auctions/useSortedAuctions';
 import { FlexContainer } from '../../../styled/Common';
 import useMediaQuery from '../../../utils/hooks/useMediaQuery';
-import useOnClickOutside from '../../../utils/hooks/useOnClickOutside';
-import AuctionLoader from '../AuctionLoader/AuctionLoader';
+
 import Chart from '../Chart/Chart';
 // import Divider from '../Divider/FutaDivider';
-import { LuCheck, LuPencil } from 'react-icons/lu';
 import { auctionDataSets } from '../../../pages/platformFuta/Account/Account';
-import Typewriter from '../TypeWriter/TypeWriter';
 import styles from './SearchableTicker.module.css';
-import TickerItem from './TickerItem';
-import { GoChevronRight } from 'react-icons/go';
+
+import useDeviceDetection from '../../../utils/hooks/useDeviceDetection';
+import ResizeableTable from '../ResizeableTable/ResizeableTable';
 
 interface propsIF {
     auctions: sortedAuctionsIF;
@@ -36,26 +29,20 @@ interface propsIF {
 }
 
 export default function SearchableTicker(props: propsIF) {
-    const { auctions, placeholderTicker, isAccount, dataState } = props;
+    const { auctions, placeholderTicker, isAccount } = props;
 
     const {
-        setIsLoading,
-        selectedTicker,
         hoveredTicker,
         setSelectedTicker,
-        setHoveredTicker,
+
         showComplete,
         setShowComplete,
         watchlists,
         setFilteredAuctionList,
     } = useContext(AuctionsContext);
 
-    const {
-        searchableTickerHeights,
-        setSearchableTickerHeight,
-        canvasRef,
-        setIsSearchableTickerHeightMinimum,
-    } = useContext(FutaSearchableTickerContext);
+    const { searchableTickerHeights, setSearchableTickerHeight, canvasRef } =
+        useContext(FutaSearchableTickerContext);
 
     const tableParentRef = useRef<HTMLDivElement>(null);
     const tickerTableRef = useRef<HTMLDivElement>(null);
@@ -80,16 +67,14 @@ export default function SearchableTicker(props: propsIF) {
         // !important:  ... adds additional data to what we already have
     }, [auctions.data.length === 0]);
 
-    const [isSortDropdownOpen, setIsSortDropdownOpen] =
-        useState<boolean>(false);
     // scrolling is disabled when hover on table
     const [isMouseEnter, setIsMouseEnter] = useState(false);
-    const customLoading = false;
 
-    const isMobile = useMediaQuery('(max-width: 1024px)');
-    const isTabletScreen = useMediaQuery(
-        '(min-width: 769px) and (max-width: 1024px)',
-    );
+    const deviceType = useDeviceDetection();
+    console.log(deviceType);
+
+    const isMobile = useMediaQuery('(max-width: 768px)');
+
     // shape of data to create filter dropdown menu options
     interface filterOptionIF {
         label: string;
@@ -118,34 +103,8 @@ export default function SearchableTicker(props: propsIF) {
         sortDropdownOptions[0],
     );
 
-    // DOM id for search input field
-    const INPUT_DOM_ID = 'ticker_auction_search_input';
-
     // variable to hold user search input from the DOM
     const [searchInputRaw, setSearchInputRaw] = useState<string>('');
-
-    // fn to clear search input and re-focus the input element
-    function clearInput(): void {
-        setSearchInputRaw('');
-        focusInput();
-    }
-
-    function focusInput(): void {
-        document.getElementById(INPUT_DOM_ID)?.focus();
-    }
-
-    // clear search input on ESC keypress
-    useEffect(() => {
-        const handleEscape = (e: KeyboardEvent): void => {
-            if (document.activeElement?.id === INPUT_DOM_ID) {
-                e.code === 'Escape' && clearInput();
-            }
-        };
-        document.body.addEventListener('keydown', handleEscape);
-        return function cleanUp() {
-            document.body.removeEventListener('keydown', handleEscape);
-        };
-    });
 
     // split auction data into complete vs incomplete subsets
     // runs any time new data is received by the components
@@ -212,16 +171,6 @@ export default function SearchableTicker(props: propsIF) {
         watchlists.shouldDisplay,
     ]);
 
-    const timeDropdownRef = useRef<HTMLDivElement>(null);
-
-    // fn to close the sort filter menu when user clicks outside
-    const clickOutsideHandler = (): void => {
-        setIsSortDropdownOpen(false);
-    };
-
-    // hook to handle user click outside of the dropdown modal
-    useOnClickOutside(timeDropdownRef, clickOutsideHandler);
-
     useEffect(() => {
         if (placeholderTicker) setSelectedTicker(undefined);
     }, [placeholderTicker]);
@@ -251,188 +200,6 @@ export default function SearchableTicker(props: propsIF) {
         }
     }, [hoveredTicker, isMouseEnter]);
 
-    // @Junior: do we still need this? `customLoading` is hardcoded to `false`
-    // @Junior: ... so this isn't actually being consumed by the app
-    if (customLoading) return <AuctionLoader setIsLoading={setIsLoading} />;
-
-    // fn to determine directionality sort arrows should indicate
-    function getArrowDirection(): 'up' | 'down' | null {
-        let output: 'up' | 'down' | null = null;
-        if (auctions.active === 'marketCap') {
-            output = auctions.isReversed ? 'up' : 'down';
-        } else if (auctions.active === 'timeLeft') {
-            output = auctions.isReversed ? 'down' : 'up';
-        }
-        return output;
-    }
-
-    // apply a consistent size to all icons inside buttons
-    const BUTTON_ICON_SIZE = 17;
-
-    const headerDisplay = (
-        <search className={styles.header}>
-            <div className={styles.search_and_sort}>
-                <div className={styles.text_search_box}>
-                    <BiSearch
-                        size={20}
-                        color='var(--text2)'
-                        id='searchable_ticker_input'
-                        onClick={() => focusInput()}
-                    />
-                    <input
-                        type='text'
-                        id={INPUT_DOM_ID}
-                        value={searchInputRaw}
-                        onChange={(e) => setSearchInputRaw(e.target.value)}
-                        placeholder='SEARCH...'
-                        spellCheck={false}
-                        autoComplete='off'
-                        tabIndex={1}
-                    />
-                    {searchInputRaw && (
-                        <MdClose
-                            size={20}
-                            color='var(--text2)'
-                            onClick={() => clearInput()}
-                        />
-                    )}
-                </div>
-                <div className={styles.sort_clickable} ref={timeDropdownRef}>
-                    <div className={styles.sort_selection}>
-                        <div
-                            className={styles.open_dropdown_clickable}
-                            onClick={() =>
-                                setIsSortDropdownOpen(!isSortDropdownOpen)
-                            }
-                        >
-                            <p>{activeSortOption.label}</p>
-                            {isSortDropdownOpen ? (
-                                <IoIosArrowUp color='var(--accent1)' />
-                            ) : (
-                                <IoIosArrowDown
-                                    onClick={(e) => {
-                                        // without this handler the click will not register
-                                        // ... properly to open the dropdown menu
-                                        e.stopPropagation();
-                                        setIsSortDropdownOpen(
-                                            !isSortDropdownOpen,
-                                        );
-                                    }}
-                                    color='var(--accent1)'
-                                />
-                            )}
-                        </div>
-                        <div
-                            className={styles.sort_direction}
-                            onClick={() => auctions.reverse()}
-                        >
-                            <IoIosArrowUp
-                                size={14}
-                                color={
-                                    getArrowDirection() === 'up'
-                                        ? 'var(--accent1)'
-                                        : ''
-                                }
-                            />
-
-                            <IoIosArrowDown
-                                size={14}
-                                color={
-                                    getArrowDirection() === 'down'
-                                        ? 'var(--accent1)'
-                                        : ''
-                                }
-                            />
-                        </div>
-                    </div>
-                    {isSortDropdownOpen && (
-                        <div className={styles.dropdown}>
-                            {sortDropdownOptions.map((item, idx) => (
-                                <p
-                                    key={idx}
-                                    className={
-                                        styles[
-                                            activeSortOption.slug === item.slug
-                                                ? 'active_sort'
-                                                : 'inactive_sort'
-                                        ]
-                                    }
-                                    onClick={() => {
-                                        setActiveSortOption(item);
-                                        setIsSortDropdownOpen(false);
-                                        auctions.update(
-                                            item.slug as auctionSorts,
-                                        );
-                                    }}
-                                >
-                                    {item.label}
-                                </p>
-                            ))}
-                        </div>
-                    )}
-                </div>
-            </div>
-            <div className={styles.filters}>
-                <button
-                    onClick={() => setShowComplete(!showComplete)}
-                    className={
-                        styles[showComplete ? 'button_on' : 'button_off']
-                    }
-                >
-                    <LuCheck size={BUTTON_ICON_SIZE} />
-                    <div>COMPLETED</div>
-                </button>
-                {isAccount || (
-                    <button
-                        onClick={() => watchlists.toggle()}
-                        className={
-                            styles[
-                                watchlists.shouldDisplay
-                                    ? 'button_on'
-                                    : 'button_off'
-                            ]
-                        }
-                    >
-                        <FaEye size={BUTTON_ICON_SIZE} />
-                        <div>WATCHLIST</div>
-                    </button>
-                )}
-                {isAccount ? (
-                    <button
-                        className={
-                            styles[
-                                dataState?.active === 'created'
-                                    ? 'button_on'
-                                    : 'button_off'
-                            ]
-                        }
-                        onClick={() => dataState?.toggle && dataState.toggle()}
-                    >
-                        <LuPencil size={BUTTON_ICON_SIZE} />
-                        <div>CREATED</div>
-                    </button>
-                ) : null}
-            </div>
-        </search>
-    );
-
-    const noAuctionsContent = (
-        <div className={styles.no_auctions_content}>
-            <Typewriter
-                text={
-                    watchlists.shouldDisplay
-                        ? 'No tickers found in your watchlist'
-                        : 'No tickers to display'
-                }
-            />
-            {watchlists.shouldDisplay && <p>Consider viewing all tickers</p>}
-            {watchlists.shouldDisplay && (
-                <button onClick={() => watchlists.toggle()}>
-                    View all tickers
-                </button>
-            )}
-        </div>
-    );
     const containerRef = useRef<HTMLDivElement>(null);
 
     useEffect(() => {
@@ -464,136 +231,6 @@ export default function SearchableTicker(props: propsIF) {
         };
     }, []);
 
-    const searchableContent = (
-        <div className={styles.ticker_table} ref={tickerTableRef}>
-            {filteredData.length ? (
-                <header>
-                    <p className={styles.cell_left}>
-                        {
-                            // this icon is a stupid but effective way
-                            // ... way to keep the header text aligned
-                            // ... with the content below
-                        }
-                        {isMobile || (
-                            <GoChevronRight
-                                size={20}
-                                className={styles.ticker_col_header_spacer}
-                            />
-                        )}
-                        TICKER
-                    </p>
-                    <p className={styles.cell_right}>MARKET CAP</p>
-                    <p className={styles.cell_center}>STATUS</p>
-                    <p className={styles.cell_right}>TIME</p>
-                    {dataState?.active === 'created' && (
-                        <p className={styles.cell_right}>ETH Committed</p>
-                    )}
-                    {dataState?.active === 'created' && (
-                        <p className={styles.cell_right}>ETH Rewards</p>
-                    )}
-                </header>
-            ) : null}
-            <div
-                className={styles.ticker_table_content}
-                onMouseEnter={() => setIsMouseEnter(true)}
-                onMouseLeave={() => {
-                    setIsMouseEnter(false);
-                    setHoveredTicker(undefined);
-                }}
-                ref={containerRef}
-            >
-                {filteredData.length
-                    ? (showComplete && auctions.active === 'timeLeft'
-                          ? [...filteredData].reverse()
-                          : [...filteredData]
-                      ).map((auction: AuctionDataIF) => (
-                          <TickerItem
-                              key={JSON.stringify(auction)}
-                              auction={auction}
-                              isAccount={isAccount}
-                              isMobile={isMobile}
-                              selectedTicker={selectedTicker}
-                              setSelectedTicker={setSelectedTicker}
-                              setShowComplete={setShowComplete}
-                              useRefTicker={tickerItemRefs}
-                              isCreated={dataState?.active === 'created'}
-                          />
-                      ))
-                    : noAuctionsContent}
-            </div>
-        </div>
-    );
-
-    const resizableChart = (
-        <Resizable
-            enable={{
-                bottom: !isAccount,
-                top: false,
-                left: false,
-                topLeft: false,
-                bottomLeft: false,
-                right: false,
-                topRight: false,
-                bottomRight: false,
-            }}
-            size={{
-                height:
-                    !isMobile || isTabletScreen
-                        ? isAccount && tableParentRef.current
-                            ? !isMobile
-                                ? tableParentRef.current.getBoundingClientRect()
-                                      .height * 0.99
-                                : '90%'
-                            : searchableTickerHeights.current
-                        : '90%',
-            }}
-            minHeight={200}
-            maxHeight={
-                isAccount || (isMobile && !isTabletScreen)
-                    ? undefined
-                    : window.innerHeight - 200
-            }
-            onResize={(
-                evt: MouseEvent | TouchEvent,
-                dir: Direction,
-                ref: HTMLElement,
-                d: NumberSize,
-            ) => {
-                const newHeight = searchableTickerHeights.current + d.height;
-                if (newHeight <= searchableTickerHeights.min) {
-                    setIsSearchableTickerHeightMinimum(true);
-                } else {
-                    setIsSearchableTickerHeightMinimum(false);
-                }
-            }}
-            onResizeStart={() => {
-                /* may be useful later */
-            }}
-            onResizeStop={(
-                evt: MouseEvent | TouchEvent,
-                dir: Direction,
-                ref: HTMLElement,
-                d: NumberSize,
-            ) => {
-                const newHeight = Math.max(
-                    searchableTickerHeights.min,
-                    Math.min(
-                        searchableTickerHeights.current + d.height,
-                        searchableTickerHeights.max,
-                    ),
-                );
-
-                setSearchableTickerHeight(newHeight);
-            }}
-            bounds={'parent'}
-            className={styles.resizable}
-            style={{ borderBottom: 'none !important' }}
-        >
-            {headerDisplay}
-            {searchableContent}
-        </Resizable>
-    );
-
     return (
         <div
             className={styles.searchable_ticker}
@@ -608,7 +245,19 @@ export default function SearchableTicker(props: propsIF) {
                 fullHeight
                 ref={tableParentRef}
             >
-                {resizableChart}
+                {/* {resizableTable} */}
+                <ResizeableTable
+                    auctions={auctions}
+                    filteredData={filteredData}
+                    isMouseEnter={isMouseEnter}
+                    setIsMouseEnter={setIsMouseEnter}
+                    containerRef={containerRef}
+                    tickerItemRefs={tickerItemRefs}
+                    searchInputRaw={searchInputRaw}
+                    setSearchInputRaw={setSearchInputRaw}
+                    activeSortOption={activeSortOption}
+                    setActiveSortOption={setActiveSortOption}
+                />
                 {!isAccount && !isMobile && <Chart />}
             </FlexContainer>
         </div>
